Add a lock to user token refresh

+++ classes/patreon_wordpress.php
@@ -237,26 +237,43 @@ class Patreon_Wordpress
     {
         $refresh_token = get_user_meta($user->ID, 'patreon_refresh_token', true);
 
-        $oauth_client = new Patreon_Oauth();
-        $refresh_data = $oauth_client->refresh_token($refresh_token, site_url().'/patreon-authorization/', false);
-
-        if (isset($refresh_data['access_token'])) {
-            update_user_meta($user->ID, 'patreon_refresh_token', $refresh_data['refresh_token']);
-            update_user_meta($user->ID, 'patreon_access_token', $refresh_data['access_token']);
-
-            return $refresh_data['access_token'];
+        if (!refresh_token || !$user->ID) {
+            return false;
         }
 
-        if (isset($refresh_data['http_status_code']) && 401 == $refresh_data['http_status_code']) {
-            // Token refresh failed, most likely invalid token data
-            delete_user_meta($user->ID, 'patreon_access_token');
-            delete_user_meta($user->ID, 'patreon_refresh_token');
-            delete_user_meta($user->ID, 'patreon_token_expires_in');
-            // TODO: Might need to consider asking the user to re-auth with
-            // Patreon.
+        $lock_key = 'patreon-wordpress-app-user-token-refresh-lock-'.$user->ID;
+
+        if (get_transient($lock_key)) {
+            return false;
         }
 
-        return false;
+        // Ensure that only one request at a time refreshes tokens for this user
+        set_transient($lock_key, true, 120);
+
+        try {
+            $oauth_client = new Patreon_Oauth();
+            $refresh_data = $oauth_client->refresh_token($refresh_token, site_url().'/patreon-authorization/', false);
+
+            if (isset($refresh_data['access_token'])) {
+                update_user_meta($user->ID, 'patreon_refresh_token', $refresh_data['refresh_token']);
+                update_user_meta($user->ID, 'patreon_access_token', $refresh_data['access_token']);
+
+                return $refresh_data['access_token'];
+            }
+
+            if (isset($refresh_data['http_status_code']) && 401 == $refresh_data['http_status_code']) {
+                // Token refresh failed, most likely invalid token data
+                delete_user_meta($user->ID, 'patreon_access_token');
+                delete_user_meta($user->ID, 'patreon_refresh_token');
+                delete_user_meta($user->ID, 'patreon_token_expires_in');
+                // TODO: Might need to consider asking the user to re-auth with
+                // Patreon.
+            }
+
+            return false;
+        } finally {
+            delete_transient($lock_key);
+        }
     }
 
     public static function updatePatreonUser()
@@ -277,7 +294,8 @@ class Patreon_Wordpress
 
         // If last update time is not empty and it is closer to time() than one day, dont update
         // TODO: comment out for testing
-        if (!('' == $last_update or ((time() - $last_update) > 86400))) {
+        $one_day_s = 86400;
+        if (!('' == $last_update or ((time() - $last_update) > $one_day_s))) {
             return false;
         }
